backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.12
aiofiles==24.1.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
httpx==0.28.1
slowapi==0.1.9
celery[redis]==5.4.0
redis==5.2.0
alembic==1.14.0
